requirements.txt
APScheduler==3.5.1
click==6.7
Flask==0.12.2
Flask-APScheduler==1.8.0
Flask-Auth==0.85
Flask-HTTPAuth==3.2.3
Flask-MySQLdb==0.2.0
funcsigs==1.0.2
futures==3.2.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
MySQL-python==1.2.5
mysqlclient==1.3.12
PyMySQL==0.8.0
PyMysqlDB==0.0.2
pyotp==2.2.6
python-dateutil==2.7.2
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.5
tzlocal==1.5.1
Werkzeug==0.14.1
